Performance Improvements

+++ fft_lines/settings.h
@@ -3,7 +3,7 @@
 
 extern HWND SettingsDlg;
 extern BOOL CALLBACK SettingsDlgProc(HWND hwnd, UINT Message, WPARAM wParam, LPARAM lParam);
-extern void setColor(float floatBetweenZeoAndOne, int* rgb);
+extern void setColor(unsigned int floatBetweenZeoAndOne, int* rgb);
 extern void initializeSettingsFile(HWND hwnd);
 extern void readSettings();
 extern void writeSettings();
